Updated Bullet to use floats

+++ src/io/anuke/ucore/entities/Damager.java
@@ -1,5 +1,5 @@
 package io.anuke.ucore.entities;
 
 public interface Damager{
-	int getDamage();
+	float getDamage();
 }
